fix: ItemPrice -> PriceListItem - update and remove fields
