--- schemas/app/ItemPrice.json
@@ -1,98 +1,0 @@
-{
-  "name": "ItemPrice",
-  "label": "Item Price",
-  "isChild": true,
-  "fields": [
-    {
-      "fieldname": "enabled",
-      "label": "Enabled",
-      "fieldtype": "Check",
-      "default": true,
-      "section": "Price List"
-    },
-    {
-      "fieldname": "buying",
-      "label": "Buying",
-      "fieldtype": "Check",
-      "placeholder": "Buying",
-      "default": false,
-      "section": "Price List"
-    },
-    {
-      "fieldname": "selling",
-      "label": "Selling",
-      "fieldtype": "Check",
-      "placeholder": "Selling",
-      "default": false,
-      "section": "Price List"
-    },
-    {
-      "fieldname": "party",
-      "label": "Party",
-      "placeholder": "Party",
-      "fieldtype": "Link",
-      "target": "Party",
-      "create": true,
-      "section": "Price List"
-    },
-    {
-      "fieldname": "item",
-      "label": "Item",
-      "fieldtype": "Link",
-      "target": "Item",
-      "required": true,
-      "create": true,
-      "section": "Item"
-    },
-    {
-      "fieldname": "unit",
-      "label": "Unit Type",
-      "fieldtype": "Link",
-      "target": "UOM",
-      "default": "Unit",
-      "section": "Item"
-    },
-    {
-      "fieldname": "rate",
-      "label": "Rate",
-      "fieldtype": "Currency",
-      "required": true,
-      "section": "Item"
-    },
-    {
-      "fieldname": "batch",
-      "label": "Batch",
-      "fieldtype": "Link",
-      "target": "Batch",
-      "create": true,
-      "section": "Item"
-    },
-    {
-      "fieldname": "validFrom",
-      "label": "Valid From",
-      "fieldtype": "Date",
-      "placeholder": "Valid From",
-      "section": "Validity"
-    },
-    {
-      "fieldname": "validUpto",
-      "label": "Valid Upto",
-      "fieldtype": "Date",
-      "placeholder": "Valid Upto",
-      "section": "Validity"
-    }
-  ],
-  "tableFields": ["item", "rate", "enabled"],
-  "quickEditFields": [
-    "enabled",
-    "buying",
-    "selling",
-    "party",
-    "item",
-    "unit",
-    "rate",
-    "batch",
-    "validFrom",
-    "validUpto"
-  ]
-}
