+++ schemas/schemas.ts
@@ -10,7 +10,7 @@ import GetStarted from './app/GetStarted.json';
 import InventorySettings from './app/inventory/InventorySettings.json';
 import Location from './app/inventory/Location.json';
 import PriceList from './app/PriceList.json';
-import ItemPrice from './app/ItemPrice.json';
+import PriceListItem from './app/PriceListItem.json';
 import PurchaseReceipt from './app/inventory/PurchaseReceipt.json';
 import PurchaseReceiptItem from './app/inventory/PurchaseReceiptItem.json';
 import SerialNumber from './app/inventory/SerialNumber.json';
@@ -103,7 +103,7 @@ export const appSchemas: Schema[] | SchemaStub[] = [
   PurchaseInvoiceItem as SchemaStub,
 
   PriceList as Schema,
-  ItemPrice as SchemaStub,
+  PriceListItem as SchemaStub,
 
   Tax as Schema,
   TaxDetail as Schema,
